Adds CPU usage monitoring functionality
Implements the retrieval of CPU usage data via a Prometheus client.

Introduces a CPU controller to expose CPU usage data
through an API endpoint. It also adds a CPU service to handle the logic of fetching and processing the CPU data from Prometheus.

Includes a REST client to interact with Prometheus and data
transfer objects to represent the Prometheus response.