Adds health check endpoint and Prometheus integration
Introduces a new health check endpoint that provides application status, version, and connection status to Prometheus and Node Exporter.

The prometheus configuration has been updated to a yaml format, and includes a target for the node-exporter.

The prometheus rest client has been modified to have a universal query function, which can be passed a query to execute.